feat(api): 2599-2600 - refactor by extracting class
